Make users_db_security.js use N=1 clusters only
The users_db_security.js test will not work against a multi-node
cluster because it relies on config settings being made by the
test code. Because there is no generic way of discovering the
locations of the other nodes on a dev cluster (they may be on
unexpected ports for one reason or another) it is only possible
to guarantee those settings are made on a single node.

This commit therefore forces the users_db_security.js test to run
against a single node cluster by:

 - setting the cluster variables to N=Q=R=W=1
 - excluding the test in the Makefile and running it explicitly
   with `dev/run -n 1`
 - teaching run_on_modified_server to correctly preserve the old
   config settings when nested run_on_modified_server calls are
   made

COUCHDB-2452